requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
